Import cleanup
Minor changes to what files contain and import scheme
